+++ sqlbuilder/insert.go
@@ -17,6 +17,11 @@ func NewInserter(tableName string) *Inserter {
 	return inserter
 }
 
+func (ins *Inserter) StructColumns(m interface{}, tagName string, omitColumns ...string) *Inserter {
+	columns := GetColumnsByModel(GetMapperByTagName(tagName), m, omitColumns...)
+	return ins.Columns(columns...)
+}
+
 func (ins *Inserter) Columns(columns ...string) *Inserter {
 	ins.columns = columns
 	return ins