+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 - [fix: Model 中包含忽略字段时创建 Dao 失败](https://github.com/fengjx/daox/pull/6)
 - [fix: [sqlbuilder] 分页sql offset 错误](https://github.com/fengjx/daox/pull/9)
 - [dao 采用组合模式引入 sqlx](https://github.com/fengjx/daox/pull/11)
+- [sqlbuilder 支持从 struct 解析字段](https://github.com/fengjx/daox/pull/13)